requirements.txt
discord.py>=2.3.0
pandas>=1.5.0
pytz>=2023.3
aiofiles>=23.0.0
requests>=2.31.0